requirements.txt
python-dotenv==0.19.1
pytube==11.0.1
pydub==0.25.1
openai==0.27.0

